+++ src/main/java/kitchenpos/domain/OrderTable.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 import javax.persistence.*;
 import java.util.Objects;
